// src/app/services/interfaces/api.interfaces.ts
export interface FetchInfoTramite {
    id:                          number;
    dependencia:                 string;
    unidadAdministrativa:        string;
    acronimoDependencia:         string;
    tipo:                        string;
    homoclave:                   string;
    nombre:                      string;
    descripcionCiudadana:        string;
    descripcion:                 string;
    modalidad:                   string;
    traResolucionesFavorables:   string;
    tramiteServicio:             string;
    traquienSolicita_Otro:       string;
    traConsultasChatLinea:       boolean;
    numeroRequFormato:           number;
    numeroRequNoFormato:         number;
    requiereConservarInfo:       null;
    acreditacion:                null;
    acreditacionEspecifique:     null;
    verificacion:                null;
    verificacionEspecifique:     null;
    inspeccion:                  null;
    inspeccionEspecifique:       null;
    conservarOtros:              null;
    conservarOtrosEspecifique:   null;
    suficienteCumplirRequisitos: boolean;
    metodologiaResolucion:       string;
    metodologiaAdjuntarArchivo:  string;
    tramiteRequiereInspeccion:   null;
    objetivoInspeccion:          null;
    traResolucionesFav:          number;
    traVolumenAnual:             number;
    traComentariosRespecto:      string;
    traConsultasChatLineaLigas:  string;
    criteriosresolucion:         Criteriosresolucion[];
    costos:                      Costo[];
    fechaActualizacion:          string;
    opcionesRealizarTramite:     OpcionesRealizarTramite[];
    oﬁcinasAtencion:             OﬁcinasAtencion[];
    contactos:                   Contacto[];
    requisitos:                  Requisito[];
    formatos:                    Formato[];
    fundamento:                  Fundamento[];
    plazos:                      Plazo[];
    vigencia:                    Vigencia[];
    solicita:                    Solicita[];
    conservarInformacion:        ConservarInformacion[];
    solicitudes:                 any[];
}

export interface ConservarInformacion {
    requiereConservar:  RequiereConservar;
    finesParaConservar: string;
    descripcion:        string;
}

export enum RequiereConservar {
    No = "No",
    Si = "Si",
}

export interface Contacto {
    nombre:         string;
    apellidoP:      string;
    apellidoM:      string;
    telefono:       string;
    extension:      string;
    cargo:          string;
    correo:         string;
    tipoContacto:   string;
    rolFuncionario: null | string;
}

export interface Costo {
    monto:                    string;
    moneda:                   string;
    tipoCosto:                string;
    momentoRealizarPago:      string;
    formula:                  string;
    dondeRealizarPago:        string;
    descripcionMedioPago:     string;
    rangoMontoInicial:        string;
    rangoMontoFinal:          string;
    montoOficinas:            null;
    montoOficinasEspecifique: string;
    montoBanco:               null;
    montoBancoReferencias:    string;
    montoEnLinea:             null;
    montoEnLineaReferencias:  string;
    montoTiendas:             null;
    montoTiendasReferncias:   string;
    montosOtros:              boolean;
    montosOtrosEspecifique:   string;
    montoOtroDesc:            string;
}

export interface Criteriosresolucion {
    criId:          number;
    criDescripcion: string;
    criTramiteId:   number;
    criActivo:      boolean;
    criFecha:       Date;
}

export interface Formato {
    nombre:                  string;
    identificador:           string;
    url:                     string;
    llenarLinea:             boolean;
    ligaLlenar:              null;
    enviarFormatos:          boolean;
    ligaEnvioFormatos:       string;
    ligaDOF:                 string;
    formatoId:               number;
    fechaPublicacionFormato: null;
}

export interface Fundamento {
    tramiteId:          number;
    tipo:               string;
    requisitoId:        number | null;
    canalAtencionId:    number | null;
    formatoId:          number | null;
    plazoPrevencionId:  null;
    plazoInterezadoId:  null;
    plazoMaximoId:      number | null;
    ordenGobierno:      OrdenGobierno;
    tipoOrdenamiento:   TipoOrdenamiento;
    nombreOrdenamiento: NombreOrdenamiento;
    articulo:           string;
    fraccion:           null | string;
    inciso:             null | string;
    parrafo:            null | string;
    numero:             null | string;
    letra:              null | string;
    otro:               null | string;
}

export enum NombreOrdenamiento {
    CódigoCivilParaElEstadoDeSANLuisPotosí = "Código Civil para el Estado de San Luis Potosí",
    LeyDeHaciendaParaElEstadoDeSANLuisPotosi = "Ley de Hacienda para el Estado de San Luis Potosi",
    LeyDelRegistroCivilDelEstadoDeSANLuisPotosí = "Ley del Registro Civil del Estado de San Luis Potosí",
    ReglamentoDeLaLeyDelRegistroCivilDelEstado = "Reglamento de la Ley del Registro Civil del Estado",
    ReglamentoInteriorDeLaSecretaríaGeneralDeGobierno = "Reglamento Interior de la Secretaría General de Gobierno",
}

export enum OrdenGobierno {
    Estatal = "Estatal",
}

export enum TipoOrdenamiento {
    Código = "Código",
    Ley = "Ley",
    Reglamento = "Reglamento",
}

export interface OpcionesRealizarTramite {
    opcionRealizarTramite: string;
    permiteAgendarCita:    boolean;
    agendarCitaEnLinea:    boolean;
    ligaCitaEnLinea:       string;
    canalAtencionId:       number;
    existeAppTramite:      null;
    accionesApp:           null;
    existeWebTramite:      null;
    ligaWeb:               string;
    accionesWeb:           null;
    viaTelefonica:         null;
    telefono:              string;
    accionesTelefono:      null;
    viaSMS:                null;
    codigoSMS:             string;
    viaKiosko:             null;
    accionesKiosko:        null;
    accionesSMS:           null;
    otro:                  null;
    especifiqueOtro:       string;
    accionesOtro:          null;
    accionesApplst:        any[];
    accionesKioskolst:     any[];
    accionesOtrolst:       any[];
    accionesPresenciallst: AccionesPresenciallst[];
    accionesSMSlst:        any[];
    accionesTelefonolst:   any[];
    accionesWeblst:        any[];
}

export interface AccionesPresenciallst {
    accion: string;
    orden:  number;
}

export interface OﬁcinasAtencion {
    id:        number;
    nombre:    string;
    direccion: string;
    telefono:  string;
    extension: null;
    horarios:  Horario[];
}

export interface Horario {
    dia:      string;
    apertura: Apertura;
    cierre:   Cierre;
}

export enum Apertura {
    The04112022080000 = "04/11/2022 08:00:00",
    The09112022030000 = "09/11/2022 03:00:00",
    The18102022080000 = "18/10/2022 08:00:00",
    The22092022080000 = "22/09/2022 08:00:00",
}

export enum Cierre {
    The04112022030000 = "04/11/2022 03:00:00",
    The09112022060000 = "09/11/2022 06:00:00",
    The18102022030000 = "18/10/2022 03:00:00",
    The22092022030000 = "22/09/2022 03:00:00",
}

export interface Plazo {
    nombrePlazo:           string;
    respuestaResolver:     null | string;
    paraPrevenir:          null | string;
    responderlaPrevencion: null | string;
    respuestaDependencia:  null | string;
}

export interface Requisito {
    nombre:                        string;
    descripcion:                   string;
    original:                      Original;
    copias:                        string;
    copiasCerificadas:             string;
    naturalezaDelRequisito:        null;
    esNecesarioFirma:              boolean;
    tipoRevisionTercero_CatId:     number | null;
    tipoRevisionTercero_CatIdNom:  null | string;
    otroTipoRevisionTercero:       string;
    empresaEmiteRevision_CatId:    number | null;
    empresaEmiteRevision_CatIdNom: null | string;
    otraEmpresaEmiteRevision:      OtraEmpresaEmiteRevision;
    requisitoEsTramite:            boolean;
    institucionEmite:              null;
    nombreTramiteRequisito:        null;
    requisitoId:                   number;
    requisitoParteFormato:         RequiereConservar;
    tiempoDias:                    number | null;
    tiempoHoras:                   number | null;
    tiempoMinutos:                 null;
    requisitoTramite:              any[] | null;
}

export enum Original {
    Empty = "",
    Sí = "Sí",
}

export enum OtraEmpresaEmiteRevision {
    Empty = "",
    OficialDelRegistroCivil = "Oficial del Registro Civil",
}

export interface Solicita {
    quienSolicita:                   string;
    casoRealizaTramite:              string;
    descripcionVinculada:            RequiereConservar;
    descripcionRequisitoOtroTramite: RequiereConservar;
}

export interface Vigencia {
    vigenciaTramite: string;
}